refactor: move handlers to separate file
