+++ go.mod
@@ -3,7 +3,7 @@ module github.com/kubescape/operator
 go 1.20
 
 require (
-	github.com/armosec/armoapi-go v0.0.254
+	github.com/armosec/armoapi-go v0.0.256
 	github.com/armosec/cluster-notifier-api-go v0.0.5
 	github.com/armosec/registryx v0.0.15
 	github.com/armosec/utils-go v0.0.40
@@ -19,7 +19,7 @@ require (
 	github.com/kubescape/backend v0.0.11
 	github.com/kubescape/go-logger v0.0.21
 	github.com/kubescape/k8s-interface v0.0.135-0.20230730135750-e6e709507847
-	github.com/kubescape/opa-utils v0.0.268
+	github.com/kubescape/opa-utils v0.0.270
 	github.com/kubescape/storage v0.0.18
 	github.com/mitchellh/mapstructure v1.5.0
 	github.com/panjf2000/ants/v2 v2.8.1
